+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.0.5-beta]
+
+- Update directory permissions after extraction completes
+- Return errors if Cleanup() fails
+- Alias `--ignore-resampled` with `-i`
+
+
 ## [0.0.4-beta]
 
 - Save file & directory permissions, timestamps, uid & gid